# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koupii LMS API",
        "description": "API documentation for English course LMS",
        "contact": {
            "email": "support@koupii.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Registers a user and logs them in",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth (Sanctum)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Referring URL Frontend for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fika Riyadi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "fika@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registered & Logged in successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Authenticates a user using email and password. Requires X-XSRF-TOKEN header if using Sanctum with session-based auth.",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth (Sanctum)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Referring URL Frontend for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "fika@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged in successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the current user",
                "description": "Invalidates the session and logs out the user. Requires X-XSRF-TOKEN header if using session-based Sanctum.",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth (Sanctum)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Referring URL Frontend for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Logout failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current authenticated user",
                "description": "Returns the authenticated user's profile data. Requires CSRF token in session-based Sanctum.",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth (Sanctum)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Referring URL Frontend for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authenticated user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Fika Riyadi"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "fika@example.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "user"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https://example.com/avatar.jpg"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "Frontend Developer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get all classes",
                "description": "Admin & Student can see all classes, Teacher can only see their own classes.",
                "operationId": "6b25f6b03203da799fad5fdc967c1457",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of classes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "20d7c432-467b-433a-b035-b43481b5ee85"
                                            },
                                            "teacher_id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Theorytical Class"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "This is theory class"
                                            },
                                            "class_code": {
                                                "type": "string",
                                                "example": "lkoiun"
                                            },
                                            "cover_image": {
                                                "type": "string",
                                                "example": "cover.jpg"
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-17T07:10:46.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-17T07:10:46.000000Z"
                                            },
                                            "teacher": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Teacher User 1"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "teacher1@example.com"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "teacher"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "bio": {
                                                        "type": "string",
                                                        "example": "Teacher account 1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/classes/create": {
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Create a new class",
                "description": "Only admin and teacher can create a new class.",
                "operationId": "b84bd7a37320a26c569ba548ef4fb038",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Theory Class"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is theory class"
                                    },
                                    "class_code": {
                                        "type": "string",
                                        "example": "ABC12345"
                                    },
                                    "cover_image": {
                                        "description": "Upload class cover image (jpg, png)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Class created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Class created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "7c8461cf-45d7-4ce3-b2c3-dfd3760783be"
                                                },
                                                "teacher_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "598d528e-c734-456e-b77c-7abee4cf92fa"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Arabic Class"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "This is arabic class"
                                                },
                                                "class_code": {
                                                    "type": "string",
                                                    "example": "123456"
                                                },
                                                "cover_image": {
                                                    "type": "string",
                                                    "example": "cover.jpg",
                                                    "nullable": true
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-17T08:44:38.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-07-17T08:44:38.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Class name already exists"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/classes/delete/{id}": {
            "delete": {
                "tags": [
                    "Classes"
                ],
                "summary": "Delete a class",
                "description": "Only admin and teacher (class owner) can delete a class.",
                "operationId": "b97ca6c1832c317ea82ce5717abd9ce1",
                "parameters": [
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "CSRF token for session-based auth",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Referer",
                        "in": "header",
                        "description": "Frontend URL for CSRF protection",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "http://localhost:3000"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Class ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Class not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}